Univariate risk premium

Risk premium for discrete risk factors in an insurance portfolio. For each level of the risk factor the risk premium is equal to the ratio between the severity and the exposure.

univariate_risk_premium(
  df,
  x,
  severity,
  exposure,
  nclaims = NULL,
  premium = NULL
)

Arguments

df
data.frame with insurance portfolio

x
column in df with risk factor

severity
column in df with severity (default is NULL)

exposure
column in df with exposure

nclaims
column in df with number of claims

premium
column in df with premium (default is NULL)

Value

An list of class univ_freq with components

df

data frame with risk premium

xvar

name of column in df with risk factor

severity

name of column in df with severity

nclaims

name of column in df with number of claims

exposure

name of column in df with exposure

premium

name of column in df with premium

Examples

univariate_risk_premium(MTPL2, x = area, severity = amount, exposure = exposure)

#>   area  amount   exposure risk_premium
#> 1    2 4063270  818.53973    4964.0474
#> 2    3 7945311  764.99178   10386.1390
#> 3    1 6896187 1065.74795    6470.7486
#> 4    0    6922   13.30685     520.1832