Automatically create a ggplot for objects obtained from fit_truncated_dist()

Takes an object produced by fit_truncated_dist(), and plots the available input.

# S3 method for truncated_dist
autoplot(
  object,
  geom_ecdf = c("point", "step"),
  xlab = NULL,
  ylab = NULL,
  ylim = c(0, 1),
  xlim = NULL,
  print_title = TRUE,
  print_dig = 2,
  print_trunc = 2,
  ...
)

Arguments

object

object univariate object produced by fit_truncated_dist()

geom_ecdf

the geometric object to use display the data (point or step)

xlab

the title of the x axis

ylab

the title of the y axis

ylim

two numeric values, specifying the lower limit and the upper limit of the scale

xlim

two numeric values, specifying the left limit and the right limit of the scale

print_title

show title (default to TRUE)

print_dig

number of digits for parameters in title (default 2)

print_trunc

number of digits for truncation values to print

...

other plotting parameters to affect the plot

Value

a ggplot2 object

Author

Martin Haringa