Generalized additive model

Fits a generalized additive model (GAM) to continuous risk factors in one of the following three types of models: the number of reported claims (claim frequency), the severity of reported claims (claim severity) or the burning cost (i.e. risk premium or pure premium).

fit_gam(
  data,
  nclaims,
  x,
  exposure,
  amount = NULL,
  pure_premium = NULL,
  model = "frequency",
  round_x = NULL
)

Arguments

data
data.frame of an insurance portfolio

nclaims
column in data with number of claims

x
column in data with continuous risk factor

exposure
column in data with exposure

amount
column in data with claim amount

pure_premium
column in data with pure premium

model
choose either 'frequency', 'severity' or 'burning' (model = 'frequency' is default). See details section.

round_x
round elements in column x to multiple of round_x. This gives a speed enhancement for data containing many levels for x.

Value

A list with components

prediction

data frame with predicted values

x

name of continuous risk factor

model

either 'frequency', 'severity' or 'burning'

data

data frame with predicted values and observed values

x_obs

observations for continuous risk factor

Details

The 'frequency' specification uses a Poisson GAM for fitting the number of claims. The logarithm of the exposure is included as an offset, such that the expected number of claims is proportional to the exposure.

The 'severity' specification uses a lognormal GAM for fitting the average cost of a claim. The average cost of a claim is defined as the ratio of the claim amount and the number of claims. The number of claims is included as a weight.

The 'burning' specification uses a lognormal GAM for fitting the pure premium of a claim. The pure premium is obtained by multiplying the estimated frequency and the estimated severity of claims. The word burning cost is used here as equivalent of risk premium and pure premium.

References

Antonio, K. and Valdez, E. A. (2012). Statistical concepts of a priori and a posteriori risk classification in insurance. Advances in Statistical Analysis, 96(2):187–224. doi:10.1007/s10182-011-0152-7.

Grubinger, T., Zeileis, A., and Pfeiffer, K.-P. (2014). evtree: Evolutionary learning of globally optimal classification and regression trees in R. Journal of Statistical Software, 61(1):1–29. doi:10.18637/jss.v061.i01.

Henckaerts, R., Antonio, K., Clijsters, M. and Verbelen, R. (2018). A data driven binning strategy for the construction of insurance tariff classes. Scandinavian Actuarial Journal, 2018:8, 681-705. doi:10.1080/03461238.2018.1429300.

Wood, S.N. (2011). Fast stable restricted maximum likelihood and marginal likelihood estimation of semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1):3-36. doi:10.1111/j.1467-9868.2010.00749.x.

Examples

fit_gam(MTPL, nclaims = nclaims, x = age_policyholder, exposure = exposure)

#>            x  predicted     lwr_95    upr_95
#> 1   18.00000 0.35181424 0.23055987 0.5368378
#> 2   18.77778 0.33540601 0.22874687 0.4917977
#> 3   19.55556 0.31976620 0.22644714 0.4515421
#> 4   20.33333 0.30486513 0.22348160 0.4158855
#> 5   21.11111 0.29067878 0.21967515 0.3846323
#> 6   21.88889 0.27718888 0.21490256 0.3575280
#> 7   22.66667 0.26438271 0.20913753 0.3342213
#> 8   23.44444 0.25225245 0.20248646 0.3142496
#> 9   24.22222 0.24079404 0.19518314 0.2970634
#> 10  25.00000 0.23000607 0.18753727 0.2820922
#> 11  25.77778 0.21988862 0.17985971 0.2688262
#> 12  26.55556 0.21044203 0.17240220 0.2568752
#> 13  27.33333 0.20166568 0.16533255 0.2459833
#> 14  28.11111 0.19355696 0.15874080 0.2360093
#> 15  28.88889 0.18611053 0.15266094 0.2268893
#> 16  29.66667 0.17931774 0.14709495 0.2185993
#> 17  30.44444 0.17316607 0.14203230 0.2111244
#> 18  31.22222 0.16763904 0.13746187 0.2044410
#> 19  32.00000 0.16271614 0.13337627 0.1985101
#> 20  32.77778 0.15837289 0.12977023 0.1932799
#> 21  33.55556 0.15458076 0.12663566 0.1886926
#> 22  34.33333 0.15130749 0.12395683 0.1846930
#> 23  35.11111 0.14851730 0.12170700 0.1812335
#> 24  35.88889 0.14617113 0.11984760 0.1782764
#> 25  36.66667 0.14422672 0.11832995 0.1757911
#> 26  37.44444 0.14263898 0.11709934 0.1737489
#> 27  38.22222 0.14136047 0.11610058 0.1721161
#> 28  39.00000 0.14034188 0.11528291 0.1708479
#> 29  39.77778 0.13953260 0.11460341 0.1698845
#> 30  40.55556 0.13888153 0.11402780 0.1691524
#> 31  41.33333 0.13833810 0.11352875 0.1685690
#> 32  42.11111 0.13785336 0.11308232 0.1680506
#> 33  42.88889 0.13738109 0.11266398 0.1675208
#> 34  43.66667 0.13687908 0.11224564 0.1669186
#> 35  44.44444 0.13631041 0.11179484 0.1662020
#> 36  45.22222 0.13564434 0.11127630 0.1653487
#> 37  46.00000 0.13485712 0.11065550 0.1643519
#> 38  46.77778 0.13393253 0.10990357 0.1632151
#> 39  47.55556 0.13286217 0.10900201 0.1619452
#> 40  48.33333 0.13164512 0.10794573 0.1605477
#> 41  49.11111 0.13028733 0.10674337 0.1590243
#> 42  49.88889 0.12880076 0.10541476 0.1573749
#> 43  50.66667 0.12720240 0.10398632 0.1556017
#> 44  51.44444 0.12551289 0.10248574 0.1537139
#> 45  52.22222 0.12375517 0.10093773 0.1517306
#> 46  53.00000 0.12195323 0.09936157 0.1496815
#> 47  53.77778 0.12013095 0.09777094 0.1476046
#> 48  54.55556 0.11831105 0.09617545 0.1455414
#> 49  55.33333 0.11651430 0.09458354 0.1435301
#> 50  56.11111 0.11475892 0.09300530 0.1416006
#> 51  56.88889 0.11306025 0.09145409 0.1397709
#> 52  57.66667 0.11143056 0.08994650 0.1380462
#> 53  58.44444 0.10987907 0.08850061 0.1364218
#> 54  59.22222 0.10841211 0.08713336 0.1348873
#> 55  60.00000 0.10703348 0.08585765 0.1334321
#> 56  60.77778 0.10574473 0.08468024 0.1320491
#> 57  61.55556 0.10454558 0.08360101 0.1307374
#> 58  62.33333 0.10343442 0.08261391 0.1295022
#> 59  63.11111 0.10240867 0.08170927 0.1283519
#> 60  63.88889 0.10146525 0.08087689 0.1272947
#> 61  64.66667 0.10060087 0.08010920 0.1263342
#> 62  65.44444 0.09981240 0.07940359 0.1254668
#> 63  66.22222 0.09909709 0.07876341 0.1246801
#> 64  67.00000 0.09845268 0.07819708 0.1239552
#> 65  67.77778 0.09787758 0.07771583 0.1232699
#> 66  68.55556 0.09737081 0.07733064 0.1226044
#> 67  69.33333 0.09693200 0.07704917 0.1219457
#> 68  70.11111 0.09656123 0.07687354 0.1212910
#> 69  70.88889 0.09625890 0.07679913 0.1206495
#> 70  71.66667 0.09602556 0.07681452 0.1200412
#> 71  72.44444 0.09586164 0.07690263 0.1194947
#> 72  73.22222 0.09576726 0.07704260 0.1190428
#> 73  74.00000 0.09574200 0.07721146 0.1187198
#> 74  74.77778 0.09578474 0.07738495 0.1185594
#> 75  75.55556 0.09589341 0.07753652 0.1185963
#> 76  76.33333 0.09606495 0.07763493 0.1188701
#> 77  77.11111 0.09629515 0.07764096 0.1194312
#> 78  77.88889 0.09657869 0.07750544 0.1203457
#> 79  78.66667 0.09690911 0.07717071 0.1216961
#> 80  79.44444 0.09727895 0.07657749 0.1235767
#> 81  80.22222 0.09767984 0.07567589 0.1260818
#> 82  81.00000 0.09810277 0.07443637 0.1292937
#> 83  81.77778 0.09853833 0.07285562 0.1332746
#> 84  82.55556 0.09897696 0.07095534 0.1380649
#> 85  83.33333 0.09940937 0.06877532 0.1436885
#> 86  84.11111 0.09982681 0.06636439 0.1501617
#> 87  84.88889 0.10022143 0.06377230 0.1575031
#> 88  85.66667 0.10058662 0.06104435 0.1657429
#> 89  86.44444 0.10091726 0.05821882 0.1749313
#> 90  87.22222 0.10120995 0.05532695 0.1851440
#> 91  88.00000 0.10146309 0.05239456 0.1964853
#> 92  88.77778 0.10167700 0.04944436 0.2090878
#> 93  89.55556 0.10185384 0.04649824 0.2231096
#> 94  90.33333 0.10199742 0.04357896 0.2387270
#> 95  91.11111 0.10211231 0.04071077 0.2561220
#> 96  91.88889 0.10220337 0.03791825 0.2754750
#> 97  92.66667 0.10227608 0.03522428 0.2969655
#> 98  93.44444 0.10233610 0.03264862 0.3207694
#> 99  94.22222 0.10238870 0.03020655 0.3470587
#> 100 95.00000 0.10243817 0.02790783 0.3760084