Include reference group in regression output

Extract coefficients in terms of the original levels of the coefficients rather than the coded variables.

rating_factors(
  ...,
  model_data = NULL,
  exposure = NULL,
  exponentiate = TRUE,
  signif_stars = TRUE,
  round_exposure = 0
)

Arguments

...
glm object(s) produced by glm()

model_data
data.frame used to create glm object(s), this should only be specified in case the exposure is desired in the output, default value is NULL

exposure
column in model_data with exposure, default value is NULL

exponentiate
logical indicating whether or not to exponentiate the coefficient estimates. Defaults to TRUE.

signif_stars
show significance stars for p-values (defaults to TRUE)

round_exposure
number of digits for exposure (defaults to 0)

Value

data.frame

Details

A fitted linear model has coefficients for the contrasts of the factor terms, usually one less in number than the number of levels. This function re-expresses the coefficients in the original coding. This function is adopted from dummy.coef(). Our adoption prints a data.frame as output.

Author

Martin Haringa

Examples

library(dplyr)
df <- MTPL2 %>%
  mutate(across(c(area), as.factor)) %>%
  mutate(across(c(area), ~biggest_reference(., exposure)))

mod1 <- glm(nclaims ~ area + premium, offset = log(exposure), family = poisson(), data = df)
mod2 <- glm(nclaims ~ area, offset = log(exposure), family = poisson(), data = df)

rating_factors(mod1, mod2, model_data = df, exposure = exposure)
#> Significance levels: *** p < 0.001; ** p < 0.01;  * p < 0.05; . p < 0.1  risk_factor       level     est_mod1     est_mod2 exposure
#> 1 (Intercept) (Intercept) 0.127954 *** 0.136993 ***       NA
#> 2        area           0 0.546558     0.548563           13
#> 3        area           1 1.000000     1.000000         1066
#> 4        area           2 0.872883     0.873953          819
#> 5        area           3 1.076233     1.078260          765
#> 6     premium     premium 1.001235           NA           NA