Generate data from truncated lognormal distribution

Random generation for the truncated log normal distribution whose logarithm has mean equal to meanlog and standard deviation equal to sdlog.

Usage

rlnormt(n, meanlog, sdlog, lower, upper)

Arguments

n

number of observations

meanlog

mean of the distribution on the log scale

sdlog

standard deviation of the distribution on the log scale

lower

numeric. Observations below this threshold are not present in the sample.

upper

numeric. Observations above this threshold are not present in the sample.

Value

The length of the result is determined by n.

Author

Martin Haringa