Univariate analysis for discrete risk factors

Univariate analysis for discrete risk factors in an insurance portfolio. The following summary statistics are calculated:

frequency (i.e. number of claims / expsore)

average severity (i.e. severity / number of claims)

risk premium (i.e. severity / exposure)

loss ratio (i.e. severity / premium)

average premium (i.e. premium / exposure)

If input arguments are not specified, the summary statistics related to these arguments are ignored.

univariate(
  df,
  x,
  severity = NULL,
  nclaims = NULL,
  exposure = NULL,
  premium = NULL
)

Arguments

df
data.frame with insurance portfolio

x
column in df with risk factor

severity
column in df with severity (default is NULL)

nclaims
column in df with number of claims (default is NULL)

exposure
column in df with exposure (default is NULL)

premium
column in df with premium (default is NULL)

Value

An list of class univ_all with components

df

data frame

xvar

name of column in df with risk factor

severity

name of column in df with severity

nclaims

name of column in df with number of claims

exposure

name of column in df with exposure

premium

name of column in df with premium

Examples

univariate(MTPL2, x = area, severity = amount, nclaims = nclaims,
                 exposure = exposure, premium = premium)

#>   area  amount nclaims   exposure premium  frequency average_severity
#> 1    2 4063270      98  818.53973   51896 0.11972540         41461.94
#> 2    3 7945311     113  764.99178   49337 0.14771401         70312.49
#> 3    1 6896187     146 1065.74795   65753 0.13699299         47234.16
#> 4    0    6922       1   13.30685     902 0.07514927          6922.00
#>   risk_premium loss_ratio average_premium
#> 1    4964.0474  78.296400        63.40071
#> 2   10386.1390 161.041632        64.49350
#> 3    6470.7486 104.880188        61.69658
#> 4     520.1832   7.674058        67.78464

# The summary statistics related to premium are not calculated
univariate(MTPL2, x = area, severity = amount, nclaims = nclaims, exposure = exposure)

#>   area  amount nclaims   exposure  frequency average_severity risk_premium
#> 1    2 4063270      98  818.53973 0.11972540         41461.94    4964.0474
#> 2    3 7945311     113  764.99178 0.14771401         70312.49   10386.1390
#> 3    1 6896187     146 1065.74795 0.13699299         47234.16    6470.7486
#> 4    0    6922       1   13.30685 0.07514927          6922.00     520.1832